+++ backend/canisters/identity/impl/src/lifecycle/post_upgrade.rs
@@ -21,6 +21,7 @@ fn post_upgrade(args: Args) {
 
     canister_logger::init_with_logs(data.test_mode, errors, logs, traces);
 
+    data.user_principals.remove_dangling_auth_principal_links();
     data.originating_canisters.insert(WEBAUTHN_ORIGINATING_CANISTER);
     data.skip_captcha_whitelist.insert(WEBAUTHN_ORIGINATING_CANISTER);