+++ backend/canisters/identity/impl/src/queries/auth_principals.rs
@@ -21,6 +21,10 @@ fn auth_principals_impl(state: &RuntimeState) -> Response {
                         principal: p,
                         originating_canister: a.originating_canister,
                         is_ii_principal: a.is_ii_principal,
+                        is_current_identity: p == auth_principal,
+                        webauthn_key: a
+                            .webauthn_credential_id
+                            .and_then(|id| state.data.webauthn_keys.get(id.clone()).map(|k| k.hydrate(id))),
                         last_used: a.last_used,
                     })
                 })