+++ frontend/openchat-agent/src/services/identity/candid/types.d.ts
@@ -19,6 +19,8 @@ export type AuthPrincipalsResponse = { 'NotFound' : null } |
     'Success' : Array<
       {
         'principal' : Principal,
+        'webauthn_key' : [] | [WebAuthnKey],
+        'is_current_identity' : boolean,
         'originating_canister' : Principal,
         'is_ii_principal' : boolean,
         'last_used' : TimestampMillis,
@@ -107,6 +109,7 @@ export interface WebAuthnKey {
   'public_key' : Uint8Array | number[],
   'origin' : string,
   'cross_platform' : boolean,
+  'aaguid' : Uint8Array | number[],
   'credential_id' : Uint8Array | number[],
 }
 export interface _SERVICE {