+++ frontend/openchat-client/src/openchat.ts
@@ -984,16 +984,13 @@ export class OpenChat extends EventTarget {
     }
 
     #startBotsPoller() {
-        const enabled = localStorage.getItem("openchat_bots_enabled") === "true";
-        if (enabled) {
-            this.#botsPoller?.stop();
-            this.#botsPoller = new Poller(
-                () => this.#loadBots(),
-                BOT_UPDATE_INTERVAL,
-                BOT_UPDATE_IDLE_INTERVAL,
-                true,
-            );
-        }
+        this.#botsPoller?.stop();
+        this.#botsPoller = new Poller(
+            () => this.#loadBots(),
+            BOT_UPDATE_INTERVAL,
+            BOT_UPDATE_IDLE_INTERVAL,
+            true,
+        );
     }
 
     #startChatsPoller() {
@@ -7105,16 +7102,18 @@ export class OpenChat extends EventTarget {
         const webAuthnOrigin = this.config.webAuthnOrigin;
         if (webAuthnOrigin === undefined) throw new Error("WebAuthn origin not set");
 
-        const webAuthnIdentity = await createWebAuthnIdentity(webAuthnOrigin);
+        const [webAuthnIdentity, aaguid] = await createWebAuthnIdentity(webAuthnOrigin);
 
         // We create a temporary key so that the user doesn't have to reauthenticate via WebAuthn, we store this key
         // in IndexedDb, it is valid for 30 days (the same as the other key delegations we use).
         const tempKey = await ECDSAKeyIdentity.generate();
+        
         return await this.#finaliseWebAuthnSignin(
             tempKey,
             () => webAuthnIdentity,
             webAuthnOrigin,
             assumeIdentity,
+            aaguid,
         );
     }
 
@@ -7130,6 +7129,7 @@ export class OpenChat extends EventTarget {
             () => webAuthnIdentity.innerIdentity(),
             webAuthnOrigin,
             true,
+            undefined,
         );
     }
 
@@ -7153,6 +7153,7 @@ export class OpenChat extends EventTarget {
             () => webAuthnIdentity,
             webAuthnKey.origin,
             false,
+            undefined,
         );
     }
 
@@ -7161,6 +7162,7 @@ export class OpenChat extends EventTarget {
         webAuthnIdentityFn: () => WebAuthnIdentity,
         webAuthnOrigin: string,
         assumeIdentity: boolean,
+        aaguid: Uint8Array | undefined,
     ): Promise<[ECDSAKeyIdentity, DelegationChain, WebAuthnKey]> {
         const sessionKey = await ECDSAKeyIdentity.generate();
         const delegationChain = await DelegationChain.create(
@@ -7177,6 +7179,7 @@ export class OpenChat extends EventTarget {
             credentialId: new Uint8Array(webAuthnIdentity.rawId),
             origin: webAuthnOrigin,
             crossPlatform: webAuthnIdentity.getAuthenticatorAttachment() === "cross-platform",
+            aaguid: aaguid ?? new Uint8Array(),
         };
         if (assumeIdentity) {
             this.#webAuthnKey = webAuthnKey;
@@ -8428,7 +8431,6 @@ export class OpenChat extends EventTarget {
     ): Promise<LinkIdentitiesResponse> {
         return this.#sendRequest({
             kind: "linkIdentities",
-            userId: this.#liveState.user.userId,
             initiatorKey: initiatorKey.getKeyPair(),
             initiatorDelegation: initiatorDelegation.toJSON(),
             initiatorIsIIPrincipal,