+++ backend/canisters/identity/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Return `last_used` on each authentication principal ([#7372](https://github.com/open-chat-labs/open-chat/pull/7372))
+- Remove auth principal from user record after unlinking ([#7398](https://github.com/open-chat-labs/open-chat/pull/7398))
 
 ## [[2.0.1594](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1594-identity)] - 2025-01-31