+++ backend/canisters/identity/CHANGELOG.md
@@ -6,10 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Implement signing in with Passkeys ([#7382](https://github.com/open-chat-labs/open-chat/pull/7382))
+- Store the `aaguid` along with each WebAuthn key ([#7409](https://github.com/open-chat-labs/open-chat/pull/7409))
+
 ### Changed
 
 - Return `last_used` on each authentication principal ([#7372](https://github.com/open-chat-labs/open-chat/pull/7372))
 - Remove auth principal from user record after unlinking ([#7398](https://github.com/open-chat-labs/open-chat/pull/7398))
+- Add `is_current_identity`+ `webauthn_key` to auth principals response ([#7410](https://github.com/open-chat-labs/open-chat/pull/7410))
+
+### Fixed
+
+- Increment originating canister count when linking new identity ([#7416](https://github.com/open-chat-labs/open-chat/pull/7416))
 
 ## [[2.0.1594](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1594-identity)] - 2025-01-31